homework 3_rev
